cheri: Make the rest of fs/ compile without warning
Standard uintptr_t vs. unsigned long issues.

+++ fs/crypto/crypto.c
@@ -213,7 +213,7 @@ struct page *fscrypt_encrypt_pagecache_blocks(struct page *page,
 		}
 	}
 	SetPagePrivate(ciphertext_page);
-	set_page_private(ciphertext_page, (unsigned long)page);
+	set_page_private(ciphertext_page, (uintptr_t)page);
 	return ciphertext_page;
 }
 EXPORT_SYMBOL(fscrypt_encrypt_pagecache_blocks);
