+++ fs/ecryptfs/crypto.c
@@ -1090,14 +1090,14 @@ out:
 	return rc;
 }
 
-static unsigned long ecryptfs_get_zeroed_pages(gfp_t gfp_mask,
+static uintptr_t ecryptfs_get_zeroed_pages(gfp_t gfp_mask,
 					       unsigned int order)
 {
 	struct page *page;
 
 	page = alloc_pages(gfp_mask | __GFP_ZERO, order);
 	if (page)
-		return (unsigned long) page_address(page);
+		return (uintptr_t) page_address(page);
 	return 0;
 }
